+++ src/data/ComplexData.ts
@@ -68,7 +68,7 @@ class ComplexData<Buffer extends DefaultBufferType = DefaultBufferType> extends 
       // 自动激活模式下主动触发激活操作
       this.changeActive('actived', 'updateData')
     }
-    const promise = this.triggerMethodWithOperate(['$updateData', args, 'update', false, (target, res) => {
+    const promise = this.triggerMethodWithOperate('$updateData', args, 'update', false, (target, res) => {
       if (target === 'start') {
         this.triggerLife('beforeUpdate', this, ...args)
       } else if (target === 'success') {
@@ -82,7 +82,7 @@ class ComplexData<Buffer extends DefaultBufferType = DefaultBufferType> extends 
           args: args
         })
       }
-    }])
+    })
     return this._setPromise('update', promise)
   }
   loadUpdateData (forceInitOption?: boolean | ForceValueInitOption | ForceValue, ...args: unknown[]) {