+++ src/dictionary/CustomEdit.ts
@@ -1,13 +1,13 @@
-import DefaultEdit, { DefaultEditInitOption } from "./DefaultEdit"
+import EditData, { EditDataInitOption } from "./EditData"
 import DictionaryValue from "../lib/DictionaryValue"
 
-export interface CustomEditInitOption extends DefaultEditInitOption {
+export interface CustomEditInitOption extends EditDataInitOption {
   type: 'custom'
   option?: Record<PropertyKey, unknown>
   custom?: Record<PropertyKey, unknown>
 }
 
-class CustomEdit extends DefaultEdit{
+class CustomEdit extends EditData{
   static $name = 'CustomEdit'
   type: 'custom'
   $option: Record<PropertyKey, unknown>